객체를 만든 후 메소드를 이용했습니다.
